backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
pyhealth>=1.1.3
polars>=0.20.0

# Dev / tests
pytest>=8.0.0
pytest-asyncio>=0.24.0
